Remove `Castle.Core.Internal.CollectionExtensions`
Its five methods are dealt with as follows:

 * `Find`, `IsNullOrEmpty`: unused.

 * `FindAll`: used only once, inlined.

 * `GetContentsHashCode`, `AreEquivalent`: used to implement equality
   contract in `ProxyGenerationOptions`. Moved there, mostly unchanged.
